Make confvalue() never return None
There are three places in the code where the return value of confvalue()
is concatenated with a string, which fails if the value is None. Rather
than changing them all, I thought it would be better to make confvalue()
return an empty string if the configuration for that value is absent.