Update project.clj - Bump snakeyaml dependency version from 1.9 to 1.11 (Latest Release)
Storm is using a snakeyaml version that is now two revisions old. I have recently come across a dependency conflict when using Storm with JRuby 1.7.3. When running in ruby 1.9 mode JRuby utilizes the snakeyaml library, however the internal JRuby  version of snakeyaml 1.11. Since Storm is running an outdated version, it would make sense to propose the change in hopes of resolving this dependency conflict.