STORM-549: "topology.enable.message.timeouts" does nothing
The ability to turn off message timeouts doesn't work.
The value is a true/false value that is represented as
a Java Boolean. The check for the truthiness used the
clojure not function. The idea was, turn off message
timeouts if "topology.enable.message.timeouts" is not true.
Unfortunately this won't worth with clojure's definition of
truthiness as:

(not (Boolean. false))

evaluates to true.

This commit updates the logic to compare the Java Boolean
"topology.enable.message.timeouts" to false using the
= function which will work.