[STORM-1176] Checkpoint window evaluated/expired state
This builds on top of the stateful bolt abstractions introduced in
PR# 939 and uses it to checkpoint the last expired and last evaluated
tuple message ids. IStatefulWindowedBolt exposes the framework managed state
to the users windowed bolt so that the state of the windowing
operation can be saved. The last expired and last evaluated message ids
are saved behind the scene along with the state of the windowing operation and this is
used during recovery to avoid duplicate window evaluations. This still provides
atleast once semantics but reduces the duplicate window evaluations.