AvroGenericRecordBolt instead of SequenceFileBolt
Changing the example code and use AvroGenericRecordBolt  instead of
SequenceFileBolt…

Section “Support for Avro Files” starts by explained the use of
org.apache.storm.hdfs.bolt.AvroGenericRecordBolt for Avro files.
However, the example code uses SequenceFileBolt instead, and that
doesn’t work, obviously because the method withSchemaAsString isn’t
part of SequenceFileBolt, but part of
org.apache.storm.hdfs.bolt.AvroGenericRecordBolt instead. So, the
change is nothing but changing the name of class SequenceFileBolt and
replace it with the correct AvroGenericRecordBolt.