STORM-1730 LocalCluster#shutdown() does not terminate all storm threads/thread pools.
The thread pool that is remaining is in a static context of the
DisruptorQueue.  To avoid major changes to how this queues lifecycle is
managed the fix is to create a thread pool factory which creates daemon
threads.  The threads are now also named with the 'disruptor-flush'
prefix to be consistent with the TimerTask.